+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ---
 
+#### 16.07.2018
+
+* [Rust] Code of Kutulu - Get closest explorer (CodinGame)
+
 #### 15.07.2018
 
 * [Rust] Code of Kutulu - Collecting units into a vec (CodinGame)